Submit assignee and comment via explicit methods
This allows the assignee to be updated when the user only has the
"assign" permission on an issue, but not the "edit issue" permission.
That is, in the web interface, you can click "Assign", but can't edit
the "Assignee" field.  Also circumvents a similar issue in which the issue
can't be assigned because the assignee field isn't on the "Edit" screen.