Fix linter issues, convert to bash
update_comments was failing on Ubuntu because "==" is undefined across
platforms. shellcheck caught it, so fix the shebang line to "#!/usr/bin/env
bash" in there and edit_jira_comment and update to remove shellcheck issues in
both scripts as well.

Perl::Critic caught the static "OUTFILE" in get_jira_issue, so change that to
"my $outfile".